Getting standalone Pravega working with loading InMemoryStorage through reflection.
