+++ segmentstore/storage/bindings/src/main/java/io/pravega/segmentstore/storage/bindings/filesystem/FileSystemStorage.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.storage.impl.filesystem;
+package io.pravega.segmentstore.storage.bindings.filesystem;
 
 import com.google.common.base.Preconditions;
 import io.pravega.common.Exceptions;
@@ -341,7 +341,7 @@ public class FileSystemStorage implements Storage {
      * Concatenation as currently implemented here requires that we read the data and write it back to target file.
      * We do not make the assumption that a native operation exists as this is not a common feature supported by file
      * systems. As such, a concatenation induces an important network overhead as each byte concatenated must be
-     * read and written back when the storage is backed by a remote filesystem (through NFS).
+     * read and written back when the storage is backed by a remote io.pravega.segmentstore.storage.bindings.filesystem.filesystem (through NFS).
      *
      * This option was preferred as other option (of having one file per transaction) will result in server side
      * fragmentation and corresponding slowdown in cluster performance.
