+++ segmentstore/storage/bindings/src/main/java/io/pravega/segmentstore/storage/bindings/filesystem/FileSystemStorageConfig.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.storage.impl.filesystem;
+package io.pravega.segmentstore.storage.bindings.filesystem;
 
 import io.pravega.common.util.ConfigBuilder;
 import io.pravega.common.util.ConfigurationException;
@@ -31,7 +31,7 @@ public class FileSystemStorageConfig {
     //region Members
 
     /**
-     * Root of the Pravega owned filesystem path. All the directories/files under this path will be exclusively
+     * Root of the Pravega owned io.pravega.segmentstore.storage.bindings.filesystem.filesystem path. All the directories/files under this path will be exclusively
      * owned by Pravega.
      */
     @Getter
