+++ segmentstore/storage/bindings/src/test/java/io/pravega/segmentstore/storage/bindings/IdempotentStorageTestBase.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.storage.impl;
+package io.pravega.segmentstore.storage.bindings;
 
 import io.pravega.segmentstore.contracts.BadOffsetException;
 import io.pravega.segmentstore.contracts.StreamSegmentNotExistsException;
