+++ segmentstore/storage/bindings/src/test/java/io/pravega/segmentstore/storage/bindings/extendeds3/S3ClientWrapper.java
@@ -30,7 +30,7 @@ import java.io.InputStream;
 import lombok.Synchronized;
 
 /**
- * Client wrapper for S3JerseyClient. It uses local filesystem to implement extended S3 JAVA client APIs.
+ * Client wrapper for S3JerseyClient. It uses local io.pravega.segmentstore.storage.bindings.filesystem.filesystem to implement extended S3 JAVA client APIs.
  */
 class S3ClientWrapper extends S3JerseyClient {
