+++ segmentstore/storage/bindings/src/test/java/io/pravega/segmentstore/storage/bindings/filesystem/FileSystemIntegrationTest.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.server.host;
+package io.pravega.segmentstore.storage.bindings.filesystem;
 
 import io.pravega.common.io.FileHelpers;
 import io.pravega.segmentstore.server.store.ServiceBuilder;
@@ -16,8 +16,8 @@ import io.pravega.segmentstore.server.store.StreamSegmentStoreTestBase;
 import io.pravega.segmentstore.storage.StorageFactory;
 import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperConfig;
 import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperLogFactory;
-import io.pravega.segmentstore.storage.impl.filesystem.FileSystemStorageConfig;
-import io.pravega.segmentstore.storage.impl.filesystem.FileSystemStorageFactory;
+import io.pravega.segmentstore.storage.bindings.filesystem.FileSystemStorageConfig;
+import io.pravega.segmentstore.storage.bindings.filesystem.FileSystemStorageFactory;
 import io.pravega.segmentstore.storage.impl.rocksdb.RocksDBCacheFactory;
 import io.pravega.segmentstore.storage.impl.rocksdb.RocksDBConfig;
 import java.io.File;
