+++ settings.gradle
@@ -19,6 +19,7 @@ include 'common',
         'segmentstore:contracts',
         'segmentstore:storage',
         'segmentstore:storage:impl',
+        'segmentstore:storage:bindings',
         'segmentstore:server',
         'segmentstore:server:host',
         'controller',
